add api/auth.go
